cambio de ruta a shop
